Move code to new files interval_analysis.{cpp,h}
